#72 Fixed tests

+++ algorithms/src/test/java/uk/ac/warwick/wsbc/tools/images/filters/HatFilter_testParam.java
@@ -35,8 +35,8 @@ import uk.ac.warwick.wsbc.plugin.QuimpPluginException;
 public class HatFilter_testParam {
     private static final Logger LOGGER = LogManager
             .getLogger(HatFilter_testParam.class.getName());
-    private Double window;
-    private Double crown;
+    private Integer window;
+    private Integer crown;
     private Double sig;
     private List<Vector2d> testcase;
     private Path testfileName;
@@ -56,8 +56,8 @@ public class HatFilter_testParam {
      * @see DataLoader
      * @see HatFilter
      */
-    public HatFilter_testParam(String testFileName, Double window, Double crown,
-            Double sig) {
+    public HatFilter_testParam(String testFileName, Integer window,
+            Integer crown, Double sig) {
         this.testfileName = Paths.get(testFileName);
         this.window = window;
         this.crown = crown;
@@ -91,10 +91,10 @@ public class HatFilter_testParam {
     @Parameterized.Parameters
     public static Collection<Object[]> testFiles() {
         return Arrays.asList(new Object[][] {
-                { "src/test/resources/testData_137.dat", 23., 13., 0.3 },
-                { "src/test/resources/testData_1.dat", 23., 13., 0.3 },
-                { "src/test/resources/testData_125.dat", 23., 13., 0.3 },
-                { "src/test/resources/testData_75.dat", 23., 13., 0.3 } });
+                { "src/test/resources/testData_137.dat", 23, 13, 0.3 },
+                { "src/test/resources/testData_1.dat", 23, 13, 0.3 },
+                { "src/test/resources/testData_125.dat", 23, 13, 0.3 },
+                { "src/test/resources/testData_75.dat", 23, 13, 0.3 } });
     }
 
     /**
